Add the default class and test for the static adapters.
The default or abstract class provides the method for getting
the same JSON representation for every retrieved line. The default
test class allows to run the same tests on the adapters.

The adapters should extend the AbstractOSMLine class. The
corresponding test must be provided. It has to extend the OSMLineTest
class and only has to provide information about the concrete adapter
class to be tested and an OpenstreetMap id for a line. The superclass
will performa the information retrieval and same tests for every adapter.

Both the default class and default test must be refractored.