Add the concrete adapter and test for the EMT Palma company
These classes implement the adapter to retrieve the information
about EMT Palma bus lines and the class to test the correctness
of the information structure.

The adapter class, OSMLine (all adapter classes must be named the
same) extends the class AbstractOSMLine, so we can get the needed
JSON representation. The test class, EMTPalmaOSMLineTest, extends,
the OSMLineTest (which, at the same time, extends from the
django.test.TestCase class), and is initialized with the OSMLine
class (the class to test) and an OpenstreetMap id that represents
a line. It calls to the static_adapter_check method in the OSMLineTest
class, so all the tests in that class can be performed.

The adapter class must be refractored.