Moved the adapters to a new Django app, implemented the view and template examples
Since it can be a reusable component, the OSM line adapters have been moved to a new
app. The view operations (in a very quick and dirty way) have been implemented (at least
the retrieval of OSM data and the preview. The final storage in the database is still
missing).

Moreover, a new requirements file has been created for development environments, so it
uses SQLite instead of Postgres. Thiss allows to have less configuration issues in
development environments.

The code has been developed quick and dirty in some hours, so it really needs a LOT
of refractoring.