Add the manager to persist the data received from the OSM adapters
Given a OSMLine data structure, the OSMLineManager converts this data into models
and persists it in the database. The order of the operations is the following:
save the line information, save the station information if it is not in the database
for every station in the line, assign the stations to the line, save the nodes
information if they are not in the database, for every route save its information,
create the default streches (one stretch per route), and for every route point,
create the corresponding mode, assign the corresponding stretch and node
and save it.