Use of an external manager to persist the OSM line data
The line detail POST method, now persists the retrieved OSM data.

To do that, it instantiates the persistence manager defined in the
settings file, so different managers can be used. Anyway, the manager
must follow two simple rules: it has to implement a constructor the line
is passed to, and it has to implement a save() method to persist the
data.