Split code in two files to avoid circular references.
The models depend on the customized managers. OSMLineManager depends
on the models. Since OSMLineManager was in the same file than the rest
of the managers, it raised a circular reference.

Moreover, in the custom managers, we were querying filtering by the
wrong fields.

Bug fixes in the OSMLineManager (like wrong field names), and use of
logging.