Update types and add indexes
The previous model used Integers as data typ for OSM ids.
It worked well in the development environment (SQLite), but
it turned out to be very large integers, which didn't work
on Postgre SQL. The solution was to use BigInteger fields
instead.

Moreover, some indexes have been added in the 'longitude'
field. This will optimize the search of nearby stations.