Add fixes to take into account timezone.
This must be corrected. our current data about time tables does not 
convert the times to UTC. To fix the deviations when querying we add
a trick and add manually the time zone offset.

This is ok for testing purposes currently, but must be fixed in order
to take into account the offset when seving time tables, not when 
querying, so any time table from all over the world can be used.

We can be automatically helped getting the timezone from geonames.