Added information of distance from beginning.
Now the distance from beginning is calculated for every route point. In
the case of stations, it is applied the one of the previous node, which
is the stop of the corresponding stop area. This information is not 
stored in the database yet.

This will be used in the future to calculate partial times depending on
the mean speed of the bus.