Adds line problems field
In the OSM line adapter, the root line dict, now can contain a problems
array. In the tests, if some problem is contained in the line
information, they fail.

As an example, in the EMT Palma and TIB adapters, problems are set when
some station is not placed into a stop area.