Changes the data structure the frontend sends to create a timetable.
The identification of stretches by generating their signatures and the
classification of the times is done in the frontend (so the backend can
focus on creating the stretches, route points and timetables.

Moreover, the station ids are not passed along the time. Since the
stations and time differences between them can be found in the
signature, only the departure times are needed, so the data structure
becomes far simpler.