User tasks are cached to improve security.
When the user "creates" a task, his name is related with the created
task id and is cached. So, when a request for check the task status is
performed, we get the user name for the request (the user must be logged
in also for calling the API), we retrieve the user tasks from the cache.
If the task does not belong to the user, the API returns an Not Allowed
HTTP response.