Map optimizations and preparation for caching.
The stations are only retrieved when the zoom is higher that 12, so we
don't make calls to the API for getting the stations in large areas.

Moreover, we round the coordinates when asking to the API to the second
decimal number. This is intended to support station caching in the
backend.