Uses a view to get all the stations for a route
Previously we were using a much more complex django model's query. It
was not only inefficient, it was now querying stop repetitions (such
as in a circular line, where the first and last stops are the same).

The only drawback is to have to create the view manually, but an SQL
is provided.