Gets the route information when the list of nearby stations is loaded
The process of retrieving all the routes and and the times for the
next vehicles is quite a consuming task. When loading the list of
nearest stations, if we load all the information before creating the
template, the main page can take a long time to load. With this change
only the nearest stations' information is loaded with the template
(a much quicker operation), and then, when the page is loaded, we
retrieve the route and time information using the REST API. The time
to load all the information is similar, but at least, the user gets
a page with the nearest stations much quicker.