Removes the routes' information when querying for a set of stations
Getting all the routes and times for every station is a very time
consuming task. Up to now, we added it for every station query, for
both only one and several. When gettign a lot of them, such as in the
map, the operation lasted too long because of this.